fix(newsletter): harvest greeting name on the real bulk send path
Review of #2936 caught a 🔴: the bulk fetchSubscribers path pre-stripped
row.email to the bare address (normalizeEmail) BEFORE passing the row to
subscriberFromFirestoreRow, so parseEmailField(row.email) inside the
function always saw an already-bare address → displayName always null →
the name harvest was a no-op on every real send (it only worked on the
single --to debug path, which passes the raw field).

- fetchSubscribers: pass the RAW row (no pre-strip); key the dedup map on
  subscriber.email (still the bare address parseEmailField returns).
- Extract subscriberFromFirestoreRow into scripts/lib/
  subscriberFromFirestoreRow.mjs so it is importable by vitest — the CLI
  send-newsletter.mjs can't be imported (shebang + import.meta.glob break
  the vite/rollup parser). This makes the harvest unit-testable end-to-end.
- tests: add 3 cases driving the real subscriberFromFirestoreRow with a
  raw "Name <addr>" row → asserts bare email + harvested first name, the
  regression guard the prior hand-mirrored test was missing.
